Add linked task exception

+++ main/java/com/timetracker/timetracker/service/ClientService.java
@@ -1,23 +1,29 @@
 package com.timetracker.timetracker.service;
 
 import com.timetracker.timetracker.model.Client;
+import com.timetracker.timetracker.model.Task;
 import com.timetracker.timetracker.repository.ClientRepository;
+import com.timetracker.timetracker.repository.TaskRepository;
 import com.timetracker.timetracker.service.exceptions.ClientNotFoundException;
+import com.timetracker.timetracker.service.exceptions.DeletedClientWithTaskException;
 import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class ClientService {
     private ClientRepository clientRepo;
+    private TaskRepository taskRepo;
 
     private String ACCESS_DENIED_MESSAGE = "User does not have ownership of this ";
 
-    public ClientService(ClientRepository clientRepo) {
+    public ClientService(ClientRepository clientRepo, TaskRepository taskRepo) {
         this.clientRepo = clientRepo;
+        this.taskRepo = taskRepo;
     }
 
     @Transactional
@@ -36,12 +42,17 @@ public class ClientService {
 
     @Transactional
     @PreAuthorize("#ownerId == principal.id")
-    public boolean deleteClient(Long ownerId, Long clientId) throws ClientNotFoundException {
+    public boolean deleteClient(Long ownerId, Long clientId) throws ClientNotFoundException, DeletedClientWithTaskException {
         Client client = clientRepo.findById(clientId)
                 .orElseThrow( () -> new ClientNotFoundException(clientId));
 
         if (client.getOwnerId() != ownerId) {
             throw new AccessDeniedException(ACCESS_DENIED_MESSAGE + "Client");
+        }
+
+        Optional<Task> linkedTask = Optional.ofNullable(taskRepo.findFirstByClientId(clientId));
+        if (linkedTask.isPresent()) {
+            throw new DeletedClientWithTaskException();
         }
 
         clientRepo.deleteById(clientId);
